Menu code
Not much yet just the basic menu code, you can add your game so far in place of the one already there. We can also do that later once we're done the game. I'll see if I can temporarily add it in. Yeah so I'm working on character and design management. Should get the sprites finished and working tomorrow.